fix(page): Close opened page if page initialization failed
I use Pyppeteer in production for a few weeks and I notice, with the command
`ps ax | ag chrome | wc -l` that the number of Chrome processes is continuously
increasing.
I noticed that new processes are created and not cleaned when an exception is
raised during the execution of the `Page.create()` function.
It makes sense because, when an exception is raised during the execution of
`Page.create()`, there is no page returned by the function, and thus the user
of Pyppeteer can't close the page.

I propose that Pyppeteer itself, before returning the exception, try to close
the Chromium browser page when the initialization of that page failed
mid-course.

Notes:

- For `asyncio.gather`, it's important to use `return_exceptions=True`,
  otherwise `Page.create()` function will raise and catch the exception
  immediatly while there is unfinished `client.send()`.

- I change `await client.send('Page.enable'),` into
  `await client.send('Page.enable')` because the `,` seemed to be wrong.

- The diff is easier to understand when adding `?w=1` inside GitHub URL.